+++ src/cupcake/utils.py
@@ -1,10 +1,11 @@
 __author__ = "miles-smith@omrf.org"
 import gzip
-from mimetypes import guess_type
 from functools import partial
-from pathlib import Path
-from typing import Union, Optional
 from io import StringIO
+from mimetypes import guess_type
+from pathlib import Path
+from typing import Optional, Union
+
 from Bio import bgzf